refactor(analysis): extract quantify stage to stages/quantify.py
Verbatim move of calculate_quantitative and _result_to_quantitative from
deep_analysis_pipeline.py into stages/quantify.py. Both symbols re-exported
from the pipeline for backward compatibility. No logic changes.
